requirements/local.txt
asgiref==3.3.1
Django==3.1.7
django-model-utils==1.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
Unipath==1.1